massa-platform/pyproject.toml
[project]
name = "massa-platform"
version = "0.1.0"
description = "Financial Intelligence Platform — RAG + Structured Data + MCP"
requires-python = ">=3.11"

dependencies = [
    "asyncpg>=0.29.0",          # Async PostgreSQL driver
    "pydantic-settings>=2.2.0", # Config from .env with type validation
    "pydantic>=2.6.0",          # Data validation
    "openai>=1.30.0",           # OpenAI embeddings API
    "voyageai>=0.2.3",          # Voyage AI embeddings (finance-domain models)
    "numpy>=1.26.0",            # Vector math for similarity tests
    "pgvector>=0.3.0",          # Registers asyncpg codec for the vector type
    "pdfplumber>=0.11.0",       # PDF text extraction (preserves layout)
    "python-docx>=1.1.0",       # Word document parsing
    "openpyxl>=3.1.0",          # Excel workbook parsing
    "tiktoken>=0.7.0",          # OpenAI tokenizer — accurate token counting
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",            # Test runner
    "pytest-asyncio>=0.23.0",   # Async test support
    "pytest-env>=1.1.0",        # Set env vars in pytest
    "fpdf2>=2.7.0",             # Creates PDF fixture files in tests
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
env = [
    "ENV=test",
]
